Add local sql db connection

+++ PizzaDelivery/Models/PizzaOrder.cs
@@ -4,7 +4,7 @@ namespace PizzaDelivery.Models
 {
     public class PizzaOrder
     {
-        public BigInteger Id { get; set; }
+        public int Id { get; set; }
         public string PizzaName { get; set; }
         public float BasePrice { get; set; }
     }
